desktop/requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
thefuzz
llama-cpp-python
huggingface-hub
fastapi
uvicorn
python-multipart
